test(gateway): isolate compaction hot-reload e2e coverage
